comply with autograder
test.cpp was a file with hashtable code which had complied with the autograder. I wanted to check what I was doing wrong in my inputs (the autograder was not passing a single test) compared to running the inputs with the test file.

The issue was with the autograder: it checks load factor >= 0.8, not just > 0.8 as specified in the question, which my code originally performed.

If this extra file of code is noticed in the version control, please excuse it. I have not copied any code from it as you can tell from my commit history, it was simply to check why the autograder was not grading me correctly when it should have.

Thank you